feat: Implement configuration handlers for %llm_config magic command
- Added ModelSetupHandler to manage model-related arguments.
- Introduced OverrideConfigHandler for parameter overrides.
- Created PersistenceConfigHandler to handle session persistence.
- Developed PersonaConfigHandler for persona-related configurations.
- Implemented SnippetConfigHandler for managing snippets.
- Added StatusDisplayHandler to show current status information.
- Refactored unit tests for the config_magic module to cover new handlers.
